+++ src/tlc_ultralytics/pose/validator.py
@@ -58,7 +58,7 @@ class TLCPoseValidator(TLCValidatorMixin, PoseValidator):
             self.sigma = np.array(self.data.get("oks_sigmas"))
 
     def postprocess(self, preds):
-        self._curr_raw_preds = preds if self._settings.collect_loss else None
+        self._stash_raw_preds(preds)
         return super().postprocess(preds)
 
     def _get_metrics_schemas(self) -> dict[str, tlc.Schema]: